--- Programacion_H2_2ºT_Pablo_Martin_Tomasi - copia/index.php
@@ -1,3 +1,0 @@
-<?php
-header('Location: ../vista/login.php');
-exit();
