+++ Programacion_H2_2ºT_Pablo_Martin_Tomasi/index.php
@@ -1,0 +1,3 @@
+<?php
+header('Location: vista/login.php');
+exit();
